fso-installer: fix installation on cygwin, prevent out of space errors
To fix the installations on Cygwin, I changed every instance of novaterm
for novacom, since it's the correct command.
For the out of space errors, I found that some commands were copying a tarball
to the Pre filesystem and then uncompressing it, this leads to out of space
problems very frequently. I changed this behaviour to send just the needed
data over novacom, leaving the uncompress process on the host computer. This
should avoid out of space errors on most situations.